requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
flask==3.0.0
gitpython==3.1.40
